Multitasking technically works.

+++ PA1/baseline.c
@@ -10,7 +10,7 @@ int main(int argc, char *argv[]){
 
     // Do Work
     int N = atoi(argv[1]);
-    int sum = 0;
+    unsigned long sum = 0;
     for(int i = 1; i <= N; i++){sum += i;} 
     
     // Record the end time
@@ -18,7 +18,7 @@ int main(int argc, char *argv[]){
 
     double time_taken = (end.tv_sec - start.tv_sec) + (double)(end.tv_nsec - start.tv_nsec) / 1000000000;
     //Print the Results
-    printf("The sum of integers from 1 to %d is: %d\n", N, sum);
+    printf("The sum of integers from 1 to %d is: %lu\n", N, sum);
     printf("This operation took %0.5f seconds.\n", time_taken);
 
 }
